apps/commentary_app/generated/requirements.txt
Flask==2.1.2
requests==2.26.0
Pillow==8.4.0
openai==0.11.3
werkzeug==2.1.2